automation/pangolin_lineage_exporter/requirements.txt
psycopg2-binary==2.9.5
webdav4[fsspec]==0.9.8
